# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Señoriales API",
    "version": "1.0.0",
    "description": "API del backend monolito de Señoriales. Cubre auth, sesiones de práctica, escenarios, integraciones de ElevenLabs y endpoints administrativos."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Desarrollo local"
    },
    {
      "url": "https://centro-de-negocios.org",
      "description": "Producción"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Invalid credentials"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/citas/directors": {
      "get": {
        "tags": [
          "Citas"
        ],
        "summary": "Lista los directores distintos presentes en las citas",
        "description": "Devuelve los nombres de director únicos, ordenados alfabéticamente. Sede-aware: un caller con scope de sede sólo ve directores de citas de asesores de su sede.",
        "responses": {
          "200": {
            "description": "Nombres de director distintos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/citas/users": {
      "get": {
        "tags": [
          "Citas"
        ],
        "summary": "Lista usuarios candidatos a asesor (para el selector de asignación)",
        "description": "Devuelve usuarios para el picker de asesor, ordenados por nombre. Sede-aware: un caller con scope de sede sólo ve usuarios de su sede.",
        "responses": {
          "200": {
            "description": "Usuarios disponibles como asesor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "email": {
                        "type": "string",
                        "format": "email"
                      },
                      "firstName": {
                        "type": "string"
                      },
                      "lastName": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/citas/{id}": {
      "patch": {
        "tags": [
          "Citas"
        ],
        "summary": "Actualiza parcialmente una cita",
        "description": "Actualiza los campos enviados de una cita existente. Sede-aware: un caller con scope de sede sólo puede modificar citas de su sede (de lo contrario devuelve 404 para no filtrar existencia). Si se reasigna el asesor, el nuevo asesor también debe ser de la misma sede.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID de la cita."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Todos los campos son opcionales (subconjunto del body de creación).",
                "properties": {
                  "director": {
                    "type": "string"
                  },
                  "asesorId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "asesorName": {
                    "type": "string"
                  },
                  "cliente": {
                    "type": "string"
                  },
                  "municipio": {
                    "type": "string"
                  },
                  "ciudad": {
                    "type": "string"
                  },
                  "zona": {
                    "type": "string"
                  },
                  "fecha": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "horaInicio": {
                    "type": "string",
                    "example": "08:00"
                  },
                  "tipo": {
                    "type": "string",
                    "enum": [
                      "presencial",
                      "virtual",
                      "telefonica"
                    ]
                  },
                  "prioridad": {
                    "type": "string",
                    "enum": [
                      "alta",
                      "media",
                      "baja"
                    ]
                  },
                  "linkSala": {
                    "type": "string"
                  },
                  "notas": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cita actualizada (incluye asesor y creator)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "director": {
                      "type": "string"
                    },
                    "asesorId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "asesorName": {
                      "type": "string"
                    },
                    "cliente": {
                      "type": "string"
                    },
                    "municipio": {
                      "type": "string"
                    },
                    "ciudad": {
                      "type": "string"
                    },
                    "zona": {
                      "type": "string"
                    },
                    "fecha": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "horaInicio": {
                      "type": "string",
                      "example": "08:00"
                    },
                    "tipo": {
                      "type": "string",
                      "enum": [
                        "presencial",
                        "virtual",
                        "telefonica"
                      ]
                    },
                    "prioridad": {
                      "type": "string",
                      "enum": [
                        "alta",
                        "media",
                        "baja"
                      ]
                    },
                    "linkSala": {
                      "type": "string",
                      "nullable": true
                    },
                    "notas": {
                      "type": "string",
                      "nullable": true
                    },
                    "createdBy": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "asesor": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "email": {
                          "type": "string",
                          "format": "email"
                        },
                        "firstName": {
                          "type": "string"
                        },
                        "lastName": {
                          "type": "string"
                        }
                      }
                    },
                    "creator": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "email": {
                          "type": "string",
                          "format": "email"
                        },
                        "firstName": {
                          "type": "string"
                        },
                        "lastName": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "El nuevo asesor no existe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "El nuevo asesor pertenece a otra sede",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Cita no encontrada (o fuera del scope de sede del caller)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Citas"
        ],
        "summary": "Elimina una cita",
        "description": "Sólo admin global o coach con acceso a la sede de la cita. La validación de sede ocurre en el service para coaches; el admin global atraviesa todas las sedes. Si la cita está fuera del scope del coach devuelve 404.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID de la cita."
          }
        ],
        "responses": {
          "200": {
            "description": "Cita eliminada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "El caller no tiene rol admin ni coach",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Cita no encontrada (o fuera del scope de sede del caller)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}